requirements.txt
pygame-ce==2.4.1
pygame-gui==0.6.9